parse xml of ecb correctly and with XmlSerializer not manually

+++ EzbAdapter/EzbAdapter/Contracts/ExchangeRate.cs
@@ -5,6 +5,6 @@ namespace EzbAdapter.Contracts
     public class ExchangeRate
     {
         public DateTime Date;
-        public float Rate;
+        public double Rate;
     }
 }
